Fix the metadata validation to not depend on the CWD
By passing a "$ref" to the validator it will figure out relative references
to other files correctly to the source file.

Since this was implemented in two places, merge them into one while at it.

getJsonSchemaStorageWithAllSchemasInABucket() is no longer used, so re-create
a config check in the config service again.